# packages/contracts/openapi.visioncart.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VisionCart API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Products",
        "operationId": "list_products_products_get",
        "parameters": [
          {
            "name": "category_slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Slug"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "brand",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand"
            }
          },
          {
            "name": "color",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Color"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductOut"
                  },
                  "title": "Response List Products Products Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Product",
        "operationId": "get_product_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/favorites": {
      "get": {
        "tags": [
          "favorites"
        ],
        "summary": "List Favorites",
        "operationId": "list_favorites_favorites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FavoriteOut"
                  },
                  "type": "array",
                  "title": "Response List Favorites Favorites Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/favorites/{product_id}": {
      "post": {
        "tags": [
          "favorites"
        ],
        "summary": "Add Favorite",
        "operationId": "add_favorite_favorites__product_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "favorites"
        ],
        "summary": "Remove Favorite",
        "operationId": "remove_favorite_favorites__product_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/guests/merge": {
      "post": {
        "tags": [
          "guests"
        ],
        "summary": "Merge Guest",
        "operationId": "merge_guest_guests_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/account": {
      "delete": {
        "tags": [
          "guests"
        ],
        "summary": "Delete Account",
        "description": "Soft-delete account: anonymize email and mark deleted_at.",
        "operationId": "delete_account_account_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeleteResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/tryon/sessions": {
      "post": {
        "tags": [
          "tryon"
        ],
        "summary": "Record Session",
        "operationId": "record_session_tryon_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TryOnSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TryOnSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/tryon/recent": {
      "get": {
        "tags": [
          "tryon"
        ],
        "summary": "Recent Tried",
        "operationId": "recent_tried_tryon_recent_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "guest_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Guest Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecentlyTriedOut"
                  },
                  "title": "Response Recent Tried Tryon Recent Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountDeleteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AccountDeleteResponse"
      },
...
      "FavoriteOut": {
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ProductOut"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "product",
          "created_at"
        ],
        "title": "FavoriteOut"
      },
...
      "GuestMergeRequest": {
        "properties": {
          "guest_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Guest Key"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "guest_key",
          "idempotency_key"
        ],
        "title": "GuestMergeRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "ProductOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "price_cents": {
            "type": "integer",
            "title": "Price Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "tryon_model_key": {
            "type": "string",
            "title": "Tryon Model Key"
          },
          "preview_url": {
            "type": "string",
            "title": "Preview Url"
          },
          "seller_url": {
            "type": "string",
            "title": "Seller Url",
            "default": ""
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "width_mm": {
            "type": "number",
            "title": "Width Mm",
            "default": 0.0
          },
          "height_mm": {
            "type": "number",
            "title": "Height Mm",
            "default": 0.0
          },
          "depth_mm": {
            "type": "number",
            "title": "Depth Mm",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "name",
          "brand",
          "color",
          "description",
          "price_cents",
          "currency",
          "tryon_model_key",
          "preview_url",
          "is_active"
        ],
        "title": "ProductOut"
      },
...
      "RecentlyTriedOut": {
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ProductOut"
          },
          "last_tried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Tried At"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          }
        },
        "type": "object",
        "required": [
          "product",
          "last_tried_at",
          "outcome"
        ],
        "title": "RecentlyTriedOut"
      },
...
      "TryOnSessionCreate": {
        "properties": {
          "session_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Session Key"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "category": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Category"
          },
          "outcome": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Outcome"
          },
          "duration_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Duration Ms",
            "default": 0
          },
          "engine": {
            "type": "string",
            "maxLength": 64,
            "title": "Engine",
            "default": ""
          },
          "device_tier": {
            "type": "string",
            "maxLength": 32,
            "title": "Device Tier",
            "default": ""
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "guest_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Key"
          }
        },
        "type": "object",
        "required": [
          "session_key",
          "product_id",
          "category",
          "outcome"
        ],
        "title": "TryOnSessionCreate"
      },
      "TryOnSessionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session_key": {
            "type": "string",
            "title": "Session Key"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_key",
          "product_id",
          "category",
          "outcome",
          "duration_ms",
          "created_at"
        ],
        "title": "TryOnSessionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}